Implement listContents, successful get some result

+++ src/B2Adapter.php
@@ -1,21 +1,31 @@
 <?php
 
-namespace H2akim\Flysystem\BackBlaze;
+namespace H2akim\Flysystem;
 
+use League\Flysystem\AdapterInterface;
+use League\Flysystem\Adapter\AbstractAdapter;
+use League\Flysystem\Util;
+use League\Flysystem\Config;
 use ChrisWhite\B2\Client;
 use ChrisWhite\B2\Bucket;
 
 class B2Adapter extends AbstractAdapter
 {
+
+    /**
+         * @var array
+         */
+        protected static $resultMap = [
+            'size' => 'size',
+            'path' => 'name',
+            'type' => 'type',
+            'id' => 'id'
+        ];
+
     /**
      * @var Client
      */
     protected $b2client;
-
-    /**
-     * @var string
-     */
-    protected $bucket_id;
 
     /**
      * @var string
@@ -43,7 +53,6 @@ class B2Adapter extends AbstractAdapter
     public function __construct(Client $client, $bucket, $prefix = '', array $options = [])
     {
         $this->b2client = $client;
-        $this->bucket_id = $client->getBucketIdFromName($bucket);
         $this->bucket_name = $bucket;
         $this->setPathPrefix($prefix);
         $this->options = $options;
@@ -57,16 +66,6 @@ class B2Adapter extends AbstractAdapter
     public function getBucketName()
     {
         return $this->bucket_name;
-    }
-
-    /**
-     * Returns bucket id
-     *
-     * @return    string
-     */
-    public function getBucketId()
-    {
-        return $this->bucket_id;
     }
 
     /**
@@ -84,38 +83,111 @@ class B2Adapter extends AbstractAdapter
     }
 
     /**
+     * Read a file
+     *
+     * @param     string $path
+     *
+     * @return    false|array false on failure file meta data on success
+     */
+    public function read($path)
+    {
+        $response = $this->b2client->getFile([
+            'FileName' => $path,
+            'BucketName' => $this->getBucketName()
+        ]);
+
+        return $response;
+
+    }
+
+    /**
      * Delete file version from bucket
      *
-     * @param
-     * @return    void
-     * @author
-     * @copyright
+     * @param     string $path
+     * @return    bool
      */
     public function delete($path)
     {
         $command = $this->b2client->deleteFile([
-            'BucketName' => getBucketName(),
-            'Filename' => $this->setPathPrefix($path)
+            'BucketName' => $this->getBucketName(),
+            'FileName' => $path
         ]);
 
         return ! $this->has($path);
     }
 
+
+
     /**
-     * Delete directory from bucket
+     * list contents
+     *
+     */
+    public function listContents($directory = '', $recursive = false)
+    {
+
+        $listing = [];
+
+        $responses = $this->b2client->listFiles([
+            'BucketName' => $this->getBucketName()
+        ]);
+
+        foreach ($responses as $response) {
+
+            $protected_property = [
+                'name' => $response->getName(),
+                'id' => $response->getId(),
+                'size' => $response->getSize(),
+                'type' => $response->getType()
+            ];
+
+            $listing[] = $this->normalizeResponse($protected_property);
+
+            if ($recursive) {
+                $listing = array_merge($listing, $this->listContents($path, true));
+            }
+        }
+
+        print_r($listing);
+
+        return $listing;
+    }
+
+    /**
+     * Normalize object result array
      *
      * @param
      * @return    void
      * @author
      * @copyright
      */
+    protected function normalizeResponse(array $response, $path = null)
+    {
+        $result = ['path' => $response['name']];
+
+        $result = array_merge($result, Util::map($response, static::$resultMap));
+
+        return $result;
+    }
+
+
+
+    /**
+     * Delete directory from bucket
+     *
+     * @param     string $path
+     * @return    bool
+     */
     public function deleteDir($path)
     {
 
-        $this->b2client->deleteFile([
-            'BucketName' => getBucketName(),
-            'Filename' => $this->setPathPrefix($path) . '/'
-        ]);
+        if ($this->has($path)) {
+
+            $this->b2client->deleteFile([
+                'BucketName' => $this->getBucketName(),
+                'FileName' => $this->setPathPrefix($path) . '/'
+            ]);
+
+        }
 
         return ! $this->has($path);
     }
@@ -123,10 +195,9 @@ class B2Adapter extends AbstractAdapter
     /**
      * Create directory in bucket (file representation)
      *
-     * @param
-     * @return    void
-     * @author
-     * @copyright
+     * @param     string $dirname
+     * @param     Config $config
+     * @return    json
      */
     public function createDir($dirname, Config $config)
     {
@@ -136,16 +207,36 @@ class B2Adapter extends AbstractAdapter
     }
 
     /**
+     * write Stream
+     *
+     */
+    public function writeStream($path, $contents, Config $config)
+    {
+        return $this->write($path, $contents, $config);
+    }
+
+
+    /**
      * Check file exists or not
      *
-     * @param
-     * @return    void
-     * @author
-     * @copyright
+     * @param     string $path
+     * @return    bool
      */
     public function has($path)
     {
-        if ($this->b2client->getFile($path)) {
+
+        $settings = [
+            'BucketName' => $this->getBucketName(),
+            'FileName' => $path
+        ];
+
+        try {
+            $file = $this->b2client->getFile($settings);
+        } catch (BadJsonException $e) {
+            $file = false;
+        }
+
+        if ($file) {
             return true;
         } else {
             return false;
@@ -163,8 +254,8 @@ class B2Adapter extends AbstractAdapter
     {
 
         $command = $this->b2client->upload([
-            'BucketName' => getBucketName(),
-            'Filename' => $path,
+            'BucketName' => $this->getBucketName(),
+            'FileName' => $path,
             'Body' => $body
         ]);
 
@@ -175,18 +266,116 @@ class B2Adapter extends AbstractAdapter
     /**
      * Description of what this does.
      *
-     * @param
-     * @return    void
+     * @param     string $prefix
+     * @return
      * @author
      * @copyright
      */
 
-    protected function setPathPrefix($prefix)
+    public function setPathPrefix($prefix)
     {
         $prefix = ltrim($prefix, '/');
 
         return parent::setPathPrefix($prefix);
     }
 
+    /* These method not yet implemented */
 
+    /**
+     * update Stream
+     *
+     */
+    public function updateStream($path, $contents, Config $config)
+    {
+        // TO BE IMPLEMENT
+    }
+        /**
+     * get meta data
+     *
+     */
+    public function getMetadata($path)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * get Mimetype
+     *
+     */
+    public function getMimetype($path)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * get Timestamps
+     *
+     */
+    public function getTimestamp($path)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * get visibility
+     *
+     */
+    public function getVisibility($path)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * getSize
+     *
+     */
+    public function getSize($path)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * Copy a file
+     *
+     */
+    public function copy($path, $newpath)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * update Stream
+     *
+     */
+    public function readStream($path)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * update Stream
+     *
+     */
+    public function setVisibility($path, $visibility)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * Update file
+     *
+     */
+    public function update($path, $contents, Config $config)
+    {
+        // TO BE IMPLEMENT
+    }
+
+    /**
+     * Rename file
+     *
+     */
+    public function rename($path, $newpath)
+    {
+        // TO BE IMPLEMENT
+    }
 }
